Change opcode from uint8 to uint32 (allows longer jumps, also faster)
0 ~/h/goawk$ time perflock ./goawk 'BEGIN { for (i=0; i<100000000; i++) s+=i; print s }'
4999999950000000

real	0m2.446s
user	0m2.448s
sys	0m0.000s

0 ~/h/goawk$ time perflock ./goawk8 'BEGIN { for (i=0; i<100000000; i++) s+=i; print s }'
4999999950000000

real	0m2.533s
user	0m2.526s
sys	0m0.009s

+++ internal/bytecode/disasm.go
@@ -75,22 +75,22 @@ func (p *Program) disassembleCode(w io.Writer, code []Opcode) {
 			writeOpcodef(w, addr, "Less")
 
 		case Jump:
-			offset := int8(code[i])
+			offset := int32(code[i])
 			i++
 			writeOpcodef(w, addr, "Jump %04x", i+int(offset))
 
 		case JumpFalse:
-			offset := int8(code[i])
+			offset := int32(code[i])
 			i++
 			writeOpcodef(w, addr, "JumpFalse %04x", i+int(offset))
 
 		case JumpTrue:
-			offset := int8(code[i])
+			offset := int32(code[i])
 			i++
 			writeOpcodef(w, addr, "JumpTrue %04x", i+int(offset))
 
 		case JumpNumLess:
-			offset := int8(code[i])
+			offset := int32(code[i])
 			i++
 			writeOpcodef(w, addr, "JumpNumLess %04x", i+int(offset))
